User delete confirmation email

+++ src/main/java/com/wiktorski/mybudget/service/UserService.java
@@ -5,17 +5,23 @@ import com.wiktorski.mybudget.model.entity.User;
 import com.wiktorski.mybudget.repository.UserRepository;
 import com.wiktorski.mybudget.service.security.SecurityService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.mail.SimpleMailMessage;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.validation.constraints.Email;
 import java.util.Optional;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
-public class UserService {
+public class UserService extends AbstractBasicService {
+
     private final UserRepository userRepo;
     private final BCryptPasswordEncoder encoder;
     private final SecurityService securityService;
+    private final EmailService emailService;
 
     public void save(User user) {
         user.setPassword(encoder.encode(user.getPassword()));
@@ -34,12 +40,22 @@ public class UserService {
         return userRepo.findByEmail(email);
     }
 
-    public void deleteUser() {
-        userRepo.delete(securityService.getLoggedInUser());
+    public void deleteUser(String token) {
+        if(token.equals(getCurrentUser().getConfirmationToken())) {
+            userRepo.delete(securityService.getLoggedInUser());
+        };
+    }
+
+    public String generateConfirmationToken() {
+        User user = getCurrentUser();
+        String token = UUID.randomUUID().toString();
+        user.setConfirmationToken(token);
+        userRepo.save(user);
+        return token;
     }
 
     /******************************************************************************************/
-    //TODO TO OTHER SERVICES???
+    //TODO those methods should be in user entity as there are stored those informations
 
     public void addToBudget(float amount){
         User user = securityService.getLoggedInUser();
@@ -59,6 +75,12 @@ public class UserService {
     public void setBudget(float amount) {
         User user = securityService.getLoggedInUser();
         user.setBudget(amount);
+    }
+
+    public void sendDeleteConfirmationEmail(HttpServletRequest request) {
+        generateConfirmationToken();
+        SimpleMailMessage deleteConfirmationEmail = emailService.createDeleteConfirmationEmail(getCurrentUser(), request);
+        emailService.sendEmail(deleteConfirmationEmail);
     }
